Fixes line number parsing on elixir v1.11 | Closes #3
